Fixed win32 build.
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@107454 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/exif/exif.c
@@ -1623,7 +1623,7 @@ static void exif_iif_add_value(image_info_type *image_info, int section_index, c
 			} else {
 				info_value = &info_data->value;
 			}
-			for (index=0,vptr=value; index<length; index++,vptr+=php_tiff_bytes_per_format[format]) {
+			for (index=0,vptr=value; index<length; index++,(int)vptr+=php_tiff_bytes_per_format[format]) {
 				if (length>1) {
 					info_value = &info_data->value.list[index];
 				}
